Move sideband and mux to hotplug module
- sideband is only used for hotplug detection
- mux *is* hotplug detection
