+++ src/lib.rs
@@ -21,11 +21,9 @@ pub mod kernel_parameters;
 pub mod logging;
 pub mod modprobe;
 pub mod module;
-pub mod mux;
 pub mod pci;
 pub mod polkit;
 pub mod radeon;
-pub mod sideband;
 pub mod snd;
 pub mod util;
 pub mod wifi;
